# i_ringer_was/i_ringer_was/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/{table_name}": {
      "get": {
        "description": "지정된 테이블의 데이터 목록을 조회합니다.",
        "operationId": "UnifiedController_findAll",
        "parameters": [
          {
            "name": "table_name",
            "required": true,
            "in": "path",
            "description": "조회할 테이블 이름",
            "schema": {
              "example": "users",
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end_date",
            "required": false,
            "in": "query",
            "description": "종료일 (YYYY-MM-DD 형식)",
            "schema": {
              "example": ""
            }
          },
          {
            "name": "start_date",
            "required": false,
            "in": "query",
            "description": "시작일 (YYYY-MM-DD 형식)",
            "schema": {
              "example": ""
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "페이지 번호 (기본값: 1)",
            "schema": {
              "example": 1
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "페이지당 항목 수 (기본값: 10)",
            "schema": {
              "example": 10
            }
          },
          {
            "name": "where",
            "required": false,
            "in": "query",
            "description": "필터링 조건을 문자열 형태로 입력 (예: status:active,type:premium)",
            "schema": {
              "example": ""
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "정렬 조건을 문자열 형태로 입력 (예: id:desc,created_at:asc)",
            "schema": {
              "example": "id:desc"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "검색 키워드",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "성공적으로 조회됨"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "데이터 목록 조회",
        "tags": [
          "통합"
        ]
      }
    },
    "/api/{table_name}/{id}": {
      "get": {
        "description": "지정된 테이블의 특정 데이터를 조회합니다.",
        "operationId": "UnifiedController_findOne",
        "parameters": [
          {
            "name": "table_name",
            "required": true,
            "in": "path",
            "description": "조회할 테이블 이름",
            "schema": {
              "example": "users",
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "조회할 레코드 ID",
            "schema": {
              "example": 1,
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "required": false,
            "in": "query",
            "description": "사용자 ID (관련 데이터 확인용)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공적으로 조회됨"
          },
          "404": {
            "description": "데이터를 찾을 수 없음"
          }
        },
        "summary": "데이터 상세 조회",
        "tags": [
          "통합"
        ]
      },
      "put": {
        "description": "지정된 테이블의 특정 데이터를 수정합니다.",
        "operationId": "UnifiedController_update",
        "parameters": [
          {
            "name": "table_name",
            "required": true,
            "in": "path",
            "description": "수정할 테이블 이름",
            "schema": {
              "example": "users",
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "수정할 레코드 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "수정할 데이터",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "성공적으로 수정됨"
          },
          "404": {
            "description": "데이터를 찾을 수 없음"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "데이터 수정",
        "tags": [
          "통합"
        ]
      },
      "delete": {
        "description": "지정된 테이블의 특정 데이터를 삭제합니다.",
        "operationId": "UnifiedController_remove",
        "parameters": [
          {
            "name": "table_name",
            "required": true,
            "in": "path",
            "description": "삭제할 테이블 이름",
            "schema": {
              "example": "users",
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "삭제할 레코드 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공적으로 삭제됨"
          },
          "404": {
            "description": "데이터를 찾을 수 없음"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "데이터 삭제",
        "tags": [
          "통합"
        ]
      }
    },
...
    "/mqtt/publish": {
      "post": {
        "operationId": "MqttController_publishMessage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "메시지 발행 성공"
          }
        },
        "summary": "특정 토픽에 메시지 발행",
        "tags": [
          "MQTT"
        ]
      }
    },
...
  },
  "info": {
    "title": "I-Ringer API",
    "description": "Hospital IV drip monitoring system API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "기본 서버"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PublishMessageDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}